feat(fire)!: campfires with fuel and smoke, and fire that spreads
- Campfire (Stone*5+Stick*3+Wood*2, hand-craft): 4-minute wood fuel with
  a 5-wood cap, warm occluded light (radius 7), 2-frame flame, smoke
  puffs that thin as fuel dies; embers relight with wood; resting within
  2 tiles doubles stamina recharge; mushrooms roast on a lit fire; fuel
  persists through saves
- FIRE SPREAD: flammable flags across wood/trees/grasses/brush; burning
  rides the tile data high bit with animated flame overlays and flicker
  light; grass fires char a few tiles and die (subcritical by tuning,
  test-locked) while wooden buildings genuinely burn down in two stages
  (walls -> plank rubble -> dirt); adjacent water/mud and rain douse it;
  standing in fire hurts; campfires spark neighbors 1-in-500 — build on
  stone, not brush (user request)
- cold-camp structures now center a burnt-out ember campfire you can
  relight
- fixes a committed usize-underflow panic in the ground-blend pass at
  negative camera coords
- 9 fire tests + night screenshot verification; includes the fishing
  wave's registry items (fish tables), which share the registry file

+++ src/gfx/lighting.rs
@@ -438,8 +438,11 @@ fn ground_blend_pass(screen: &mut Screen, g: &Game, lvl: usize, x_scroll: i32, y
     let seed = g.world_seed;
     let tx0 = x_scroll >> 4;
     let ty0 = y_scroll >> 4;
-    let nx = ((x_scroll + screen::W - 1) >> 4) as usize - tx0 as usize + 1;
-    let ny = ((y_scroll + screen::H - 1) >> 4) as usize - ty0 as usize + 1;
+    // Span math stays in i32 until the end: at negative scroll (west/north of the
+    // origin — half the infinite world) casting the floored tile coords to usize
+    // first wrapped huge and the subtraction overflowed.
+    let nx = (((x_scroll + screen::W - 1) >> 4) - tx0 + 1) as usize;
+    let ny = (((y_scroll + screen::H - 1) >> 4) - ty0 + 1) as usize;
 
     // Per-tile factors over the visible grid plus a one-tile margin (edge corners
     // average tiles just off screen).
